Travis: test against PHP 7.4
While Travis (still) doesn't have an official PHP 7.4 image, this unofficial image made available by one of the PHP core contributors is up-to-date and allows you to test against PHP 7.4.

As PHP 7.4 is supposed to be released by end of November, I'm not setting `allow_failures` as the code really should be compatible before that time.