Tests: add minimal `expectExceptionMessageMatches()` compatibility wrapper to one test file
To make the tests PHPUnit cross-version compatible, not that much is needed in terms of compatibility polyfills.

The `expectExceptionMessageRegExp()` method has been removed in favour of the `expectExceptionMessageMatches()` method in PHPUnit 9.

As that assertion is only used in one test file, adding a wrapper method which toggles between the "old" method name and the "new" method name fixed cross-version compatibility for this.