Tests: add minimal `assertStringContainsString()` compatibility wrapper to one test
To make the tests PHPUnit cross-version compatible, not that much is needed in terms of compatibility polyfills.

The PHPUnit `assertContains()` method could originally be used to check whether an array contained a certain needle or whether a string contained a substring.
This second use has been deprecated in PHPUnit 8 and removed in PHPUnit 9 in favour of dedicated string related assertions which were introduced in PHPUnit 7.5.

As that assertion is only used in one specific test, adding a simple  if/else wrapper which toggles between the "old" method name and the "new" method name fixed cross-version compatibility for this.