Docs: add x-links for WP stubs
The WordPress specific `stubEscapeFunctions()` and `stubTranslationFunctions()` tools are mentioned in the "Bulk patching with stubs()" section, but not in the WordPress specific tools section.

This attempts to make these (very useful) functions more discoverable by adding links to the "Bulk patching with stubs()" section from the "WordPress testing tools" section.