Make the `ClosureParamStringForm` class compatible with PHP 5.6
This commit does makes the `ClosureParamStringForm` class compatible with PHP 5.6, both for the `fromString()`, as well as the `fromReflectionParameter()` method.

### Regarding the `fromReflectionParameter()` method:

In PHP 5.6, to retrieve the type, typically a pattern along the following lines would be used:
```php
try {
    $hasType = $reflParameter->getClass();
} catch ( ReflectionException $e ) {
    // Class with a type declaration for a non-existent class.
}

$typeName = $hasType->name;
```

The downside of the above is that the `ReflectionParameter::getClass()` method triggers an autoload of the class (for class-based type declarations).
See: https://www.php.net/manual/en/reflectionparameter.getclass.php#108620

In a test situation, this autoload should not be triggered by the tests, so instead of this pattern, the `ReflectionParameter::__toString()` method is used to retrieve a string representation of the parameter declaration, which is subsequently parsed using a regular expression.

The tests which were originally added in 95 for PHP 5.6, have been brought back and adjusted for this manner of handling the parameter.

### Regarding the `fromString()` method:

This method was already perfectly capable of handling the parameter parsing in PHP 5.6, but the `(PHP_MAJOR_VERSION < 7) and $type_name = '';` condition in the `ClosureParamStringForm::__construct()` method prevented it from storing the retrieved `$type_name`.

Now the `fromReflectionParameter()` method has been made compatible with PHP 5.6, that line in the `ClosureParamStringForm::__construct()` method can be removed and the separate set of tests for PHP 5.6 for the `fromString()` method can be removed.

## Regarding the `ClosureStringFormTest` class

As both entry methods to the `ClosureParamStringForm` class are now compatible with PHP 5.6, the segregated tests for PHP 5.6 vs PHP 7+ in the `ClosureStringFormTest` class can be consolidated again.

Fixes 94