GH Actions: run the tests on PHP >= 7.3
I was looking through the action output and was a bit confused by the test runs on PHP >= 7.3 not showing any run information.

Turns out, the tests haven't been running on PHP > 7.3 since October last year, as the `--migrate-configuration` CLI argument is supposed to be used as a stand-alone command and does not trigger a test run after the migration has finished....

Fixed now. :fingers_crossed: the tests will still pass.