Composer: update PHPUnit version constraints [1]
While PHPUnit as of version 8.5.12 allows for installation on PHP 8.0, it still contained a [nasty bug](https://github.com/sebastianbergmann/phpunit/issues/4575) which would error out the test run with a Fatal when PHPUnit 8.5.12 would be installed for PHP 8.1.

This bug was fixed in version 8.5.14.

This adjustment of the version constraints means that when `composer install` is run with `--prefer-lowest` on PHP 8.x, PHPUnit 8.5.14 will be installed instead of PHP 8.5.12, while still allowing for all other 8.x versions in all other circumstances.

Ref: https://github.com/sebastianbergmann/phpunit/blob/8.5/ChangeLog-8.5.md