Composer: update PHPUnit version constraints [2]
PHPUnit 5.x allows for installation on PHP 5.6 and 7.x, however until version 5.7.26, it [used the `each()` function](https://github.com/sebastianbergmann/phpunit/issues/2472), which was deprecated in PHP 7.2, which means that test runs against PHPUnit < 5.7.26 in combination with PHP 7.2 and higher would error out.

This adjustment of the version constraints means that when `composer install` is run with `--prefer-lowest` on PHP 5.6/7.x, PHPUnit 5.7.26 will be installed instead of PHP 5.7.9, preventing this issue.

Ref: https://github.com/sebastianbergmann/phpunit/blob/5.7.27/ChangeLog-5.7.md