WP Helper functions: ensure parameter names are in line with WP Core
... for compatibility with code under test using PHP 8.0+ function calls with named arguments.

For some of these functions, WP has renamed the parameter(s) over the past year or so, for some, the names were never in line.

Even though WP Core doesn't necessarily officially _support_ the use of named parameters in function calls, that (of course) won't stop people from using them and their tests should not break on that due to BrainMonkey not being in line with WP for parameter names.