WP Helper functions - `has_*()`: ensure parameter defaults are in line with WP Core
The `has_filter()` and `has_action()` functions in WP allow for calling them without a `$callback` parameter, but in that case, the default value should be `false`, not `null`.

While - until now[*] - this probably wasn't problematic, the BrainMonkey functionality should still mirror the WP functionality and in this case, it didn't.

[*] Why until now ?
Because as of WP 6.9, the `has_filter()` and `has_action()` functions have gotten a third parameter and if people want to pass that, but not the second parameter, correctly handling the (real) default value of the second parameter suddenly becomes relevant.

Fixed now.

Includes tests.
Includes fixing the documentation.

Refs:
* https://developer.wordpress.org/reference/functions/has_filter/
* https://developer.wordpress.org/reference/functions/has_action/