PHP 8.4 | Fix implicitly nullable parameter
PHP 8.4 deprecates implicitly nullable parameters, i.e. typed parameter with a `null` default value, which are not explicitly declared as nullable.

This commit fixes the one instance found in this codebase.

Includes updating the documentation to match.

Note: while this is not a `final` class and the change is effectively a change to the method signature, this is not a BC break as (overloaded) constructors are exempt from signature checks.

Ref: https://wiki.php.net/rfc/deprecate-implicitly-nullable-types