GH Actions: allow test runs to succeed on fork
As things were, test runs on forks (when the `master` branch in the fork is updated) would always fail on the "upload code coverage reports" step, as forks (justifiably) don't have access to the `CODECOV_TOKEN`.

Fixed now by adding a condition for running that step and for efficiency, adding the same condition to the steps running the tests.