Composer: bump PHPUnit version constraints
In some cases, CI could currently fail with `dependencies = lowest` due to the PHPUnit version being installed not being compatible with the PHP version on which the tests are being run.

This shouldn't happen and is not a good reason for the test run to fail.

With this in mind and considering that PHPUnit is a `dev` dependency in the `composer.json` file, not a `require` dependency, I'm proposing to bump the version constraints to the latest point releases for each PHPUnit minor we need to run the tests.

Mind: this, of course, means that the tests will no longer be run against old PHPUnit versions and that some BrainMonkey functionality _may_ not work when used in combination with old PHPUnit versions.

In my opinion, this is fine as we should be able to expect users to use BrainMonkey in combination with the latest compatible PHPUnit version for the PHP version they are running the tests on.