GH Actions: move PHPCompatibility check to separate job
As things were, the PHP 7.4 build could unexpectedly fail on a PHPCompatibility issue. This violates the principle of least surprise for contributors.

This commit removes the step which runs PHPCompatibility from the `qa` workflow and sets up a separate job to run PHPCompatibility.

In addition to this, it improves how the PHPCompatibility task is run by:
1. Always running against the `latest` stable PHP version.
2. No longer being subject to the `error_reporting=-1`, which could lead to PHP deprecation notices for PHPCS/PHPCompatibility breaking the scan.
3. Telling PHPCS to create both a "full" as well as a checkstyle report.
    The "full" report will be displayed in the GHA logs.
    The "checkstyle" report will be used by `cs2pr` to provide inline annotations with feedback in PRs.
4. Turns on parallel running, which should make the task fast to run.