GH Actions: move PHP linting to separate job
Generally speaking, if a project has high test coverage and the tests are being run in CI, the loading of the files for running the tests will uncover parse/compile errors and fail the build.

However, the messages about this coming from PHPUnit may not be as clear-cut. Additionally, there may be a situation (now or later) where there are some select PHP files in the project which aren't run in the test job, so having a "lint" job is still a good idea.

This commit removes the step which runs PHP Parallel Lint from the `qa` workflow and sets up a separate job to run PHP Parallel Lint.

In addition to this, it improves how the PHP Parallel Lint task is run by:
1. Only running against the high/low of each PHP major + PHP nightly.
    That should be enough to caught all parse/compile errors.
2. Also linting the `inc` subdirectory, which was previously skipped/missed.
3. PHP can show a limited number of deprecation messages during linting. By default this ability is turned off.
    This ability has now been explicitly turned on.
4. Setting the report format to "checkstyle" and letting the `cs2pr` tool provide inline annotations about parse/compile errors found in PRs.
5. Making the `test` job dependent on the `lint` job, as we already know it will fail if the `lint` job fails, so let's not waste any resources starting up the test builds.