GH Actions: improve test workflow
The builds against PHP 8.4, 8.5 and 8.6 were so far only run against the "highest" (stable) versions of dependencies due to running the tests with `--prefer-lowest` running into trouble.

With the analysis I did for improving the Mockery version constraints in the `composer.json` file as per the previous commit, I've identified Mockery `1.6.10` as the lowest viable version to run the tests with for PHP 8.4, 8.5 and 8.6.

This commit changes the workflow to allow a certain form of running against "lowest" dependencies for PHP 8.4, 8.5, 8.6, by explicitly setting the "low" version of Mockery to `1.6.10`.